tested some lsl issues and added a log for the latest sample, updated version

+++ sksevr/RealVirtualMagic/Engine.h
@@ -26,7 +26,7 @@ namespace RealVirtualMagic
 	void HookAnimations();
 	void StartCast();
 	void EndCast();
-	void ApplyFocusValue(float newFocus);
+	void ApplyFocusValue(double newFocus);
 
 	bool RegisterFuncs(VMClassRegistry* registry);
 }
